Make banner component reusable
